+++ src/coq_elpi_builtins_HOAS.ml
@@ -41,8 +41,8 @@ type trm       term   -> argument. % Eg. (t).
 %
 % Eg. Record m A : T := K { f : t; .. }.
 type indt-decl indt-decl -> argument.
-% Eg. Definition m A : T := B. (or Axiom to omit the body, body comes first)
-type const-decl id -> option term -> option term -> argument.
+% Eg. Definition m A : T := B. (or Axiom when the body is none)
+type const-decl id -> option term -> arity -> argument.
 % Eg. Context A (b : A).
 type ctx-decl context-decl -> argument.
 
@@ -174,7 +174,7 @@ evar _ _ _. % volatile, only unresolved evars are considered as evars
 % To ease the creation of a context with decl and def
 % Eg.  @pi-decl `x` <t> x1\ @pi-def `y` <t> <v> y\ ...
 macro @pi-decl N T F :- pi x\ decl x N T => F x.
-macro @pi-def N T B F :- pi x\ def x N T B => F x.
+macro @pi-def N T B F :- pi x\ def x N T B => cache x B_ => F x.
 
 % %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % Coq's goals and tactic invocation (coq_elpi_goal_HOAS.ml)
@@ -252,47 +252,83 @@ typeabbrev opaque?   bool.  macro @opaque!   :- tt.
 typeabbrev global?   bool.  macro @global!   :- tt.
 typeabbrev local?    bool.  macro @local!    :- tt.
 
-% Declaration of inductive types
+% Declaration of inductive types %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
 kind indt-decl type.
 kind indc-decl type.
 kind record-decl type.
-type parameter   id -> term -> (term -> indt-decl) -> indt-decl.
-type inductive   id -> int -> term -> (term -> list indc-decl) -> indt-decl.
-type coinductive id -> int -> term -> (term -> list indc-decl) -> indt-decl.
-type constructor id -> term -> indc-decl.
+
+% An arity is written, in Coq syntax, as:
+%    (x : T1) .. (xn : Tn) : S1 -> ... -> Sn -> U
+% This syntax is used, for example, in the type of an inductive type or
+% in the type of constructors. We call the abstractions on the left of ":"
+% "parameters" while we call the type following the ":" (proper) arity.
+
+% Note: in some contexts, like the type of an inductive type constructor,
+% Coq makes no distinction between these two writings
+%    (xn : Tn) : forall y1 : S1, ...    and      (xn : Tn) (y1 : S1) : ...
+% while Elpi is a bit more restrictive, since it understands user directives
+% such as the implicit status of an arguments (eg, using {} instead of () around
+% the binder), only on parameters.
+% Moreover parameters carry the name given by the user as an "id", while binders
+% in terms only carry it as a "name", an irrelevant pretty pringintg hint (see
+% also the HOAS of terms). A user command can hence only use the names of
+% parameters, and not the names of "forall" quantified variables in the arity.
+%
+% See also the arity->term predicate in coq-lib.elpi
+
+type parameter id -> implicit_kind -> term -> (term -> arity) -> arity.
+type arity term -> arity.
+
+type parameter   id -> implicit_kind -> term -> (term -> indt-decl) -> indt-decl.
+type inductive   id -> bool -> arity -> (term -> list indc-decl) -> indt-decl. % tt means inductive, ff coinductive
 type record      id -> term -> id -> record-decl -> indt-decl.
+
+type constructor id -> arity -> indc-decl.
+
 type field       field-attributes -> id -> term -> (term -> record-decl) -> record-decl.
 type end-record  record-decl.
-% Eg (remark A is a parameter, y is a non-uniform parameter and t also has
-% an index of type bool):
+
+% Example.
+% Remark that A is a regular parameter; y is a non-uniform parameter and t
+% also features an index of type bool.
 %
-%  Inductive t (A : Type) (y : nat) : bool -> Type :=
-%    K1 (x : A) n (p : S n = y) (e : t A n true) : t A y true
+%  Inductive t (A : Type) | (y : nat) : bool -> Type :=
+%  | K1 (x : A) {n : nat} : S n = y -> t A n true -> t A y true
 %  | K2 : t A y false
 %
 % is written
 %
-%  (parameter `A` {{ Type }} a\
-%     inductive "t" 1 {{ nat -> bool -> Type }} t\
-%       [ constructor "K1" {{ forall y,
-%           forall (x : lp:a) n (p : S n = y) (e : lp:t n true),
-%           lp:t y true }}
-%       , constructor "K2" {{ forall y,
-%           lp:t y false }} ])
+%  (parameter "A" explicit {{ Type }} a\
+%     inductive "t" tt (parameter "y" explicit {{ nat }} _\
+%                     arity {{ bool -> Type }})
+%      t\
+%       [ constructor "K1"
+%          (parameter "y" explicit {{ nat }} y\
+%           (parameter "x" explicit a x\
+%            (parameter "n" maximal {{ nat }} n\
+%              arity {{ S lp:n = lp:y -> lp:t lp:n true -> lp:t lp:y true }})))
+%       , constructor "K2"
+%          (parameter "y" explicit {{ nat }} y\
+%            arity {{ lp:t lp:y false }}) ])
 %
-% Remark that the uniform parameters do not have to be passed to t, since
-% they never change, while non-uniform parameters have to be both abstracted
-% in each constructor type and passed as arguments to t. Inside the declaration
-% the type of t is the one given just before t\ (while in Coq the implicit
-% argument status can change what one is supposed to pass to t).
-% Finally the coq.typecheck-indt-decl can be used to fill in implicit arguments
-% an infer universe constraints in the declaration above (the quotation adds
-% an implicit argument for the type of y and for the argument to eq).
+% Remark that the uniform parameters are not passed to occurrences of t, since
+% they never change, while non-uniform parameters are both abstracted
+% in each constructor type and passed as arguments to t.
+%
+% The coq.typecheck-indt-decl API can be used to fill in implicit arguments
+% an infer universe constraints in the declaration above (e.g. the hidden
+% argument of "=" in the arity of K1).
+%
+% Note: when and inductive type declaration is passed as an argument to an
+% Elpi command non uniform parameters must be separated from the uniform ones
+% with a | (a syntax introduced in Coq 8.12 and accepted by coq-elpi since
+% version 1.4, in Coq this separator is optional, but not in Elpi).
 
 % Context declaration (used as an argument to Elpi commands)
 kind context-decl type.
 % Eg. (x : T) or (x := B), body is optional, type may be a variable
-type context-item  id -> term -> option term -> (term -> context-decl) -> context-decl.
+type context-item  id -> implicit_kind -> term -> option term -> (term -> context-decl) -> context-decl.
 type context-end   context-decl.
 
 typeabbrev field-attributes (list field-attribute).