+++ src/coq_elpi_glob_quotation.mli
@@ -9,12 +9,32 @@ open RawData
 val gterm2lp :
   depth:int -> State.t -> Glob_term.glob_constr -> State.t * term
 
+
+
 val under_ctx :
   Names.Name.t ->
   term ->
   term option ->
-  (depth:int -> State.t -> 'a -> State.t * 'b) ->
-  depth:int -> State.t -> 'a -> State.t * 'b
+  (depth:int -> State.t -> State.t * 'b) ->
+  depth:int -> State.t -> State.t * 'b
+
+val do_term :
+  Glob_term.glob_constr ->
+  depth:int -> State.t -> State.t * term
+val do_params :
+  Glob_term.glob_decl list ->
+  (depth:int -> State.t -> State.t * term) ->
+  depth:int -> State.t -> State.t * term
+val do_arity :
+  Glob_term.glob_constr ->
+  depth:int -> State.t -> State.t * term
+
+val do_record :
+  name:string list * Names.Id.t ->
+  constructorname:Names.Id.t option ->
+  Glob_term.glob_constr ->
+  (Glob_term.glob_constr * Coq_elpi_HOAS.record_field_spec) list ->
+  depth:int -> State.t -> State.t * term
 
 (* The context used to interpret Var("x") nodes *)
 val set_coq_ctx_hyps : State.t -> Coq_elpi_HOAS.coq_context * Coq_elpi_HOAS.hyp list -> State.t